+++ Generator/Source/GenerateMocksCommand.swift
@@ -12,9 +12,9 @@ import SourceKittenFramework
 import FileKit
 import CuckooGeneratorFramework
 
-private func curry<P1, P2, P3, P4, P5, P6, R>(f: (P1, P2, P3, P4, P5, P6) -> R)
-    -> (P1) -> (P2) -> (P3) -> (P4) -> (P5) -> (P6) -> R {
-        return { p1 in { p2 in { p3 in { p4 in { p5 in { p6 in f(p1, p2, p3, p4, p5, p6) } } } } } }
+private func curry<P1, P2, P3, P4, P5, P6, P7, R>(f: (P1, P2, P3, P4, P5, P6, P7) -> R)
+    -> (P1) -> (P2) -> (P3) -> (P4) -> (P5) -> (P6) -> (P7) -> R {
+        return { p1 in { p2 in { p3 in { p4 in { p5 in { p6 in { p7 in f(p1, p2, p3, p4, p5, p6, p7) } } } } } } }
 }
 
 public struct GenerateMocksCommand: CommandType {
@@ -23,7 +23,8 @@ public struct GenerateMocksCommand: CommandType {
     public let function = "Generates mock files"
     
     public func run(options: Options) -> Result<Void, CuckooGeneratorError> {
-        let parsedFiles = options.files.map { File(path: $0) }.flatMap { $0 }.map { Tokenizer(sourceFile: $0).tokenize() }
+        let tokens = options.files.map { File(path: $0) }.flatMap { $0 }.map { Tokenizer(sourceFile: $0).tokenize() }
+        let parsedFiles = options.noClassMocking ? removeClasses(tokens) : tokens
         
         let headers = parsedFiles.map { options.noHeader ? "" : FileHeaderHandler.getHeader($0, withTimestamp: !options.noTimestamp) }
         let imports = parsedFiles.map { FileHeaderHandler.getImports($0, testableFrameworks: options.testableFrameworks) }
@@ -52,6 +53,13 @@ public struct GenerateMocksCommand: CommandType {
         return .Success()
     }
     
+    private func removeClasses(filesRepresentation: [FileRepresentation]) -> [FileRepresentation] {
+        return filesRepresentation.map {
+                let declarations = $0.declarations.filter { !($0 is ClassDeclaration) }
+                return FileRepresentation(sourceFile: $0.sourceFile, declarations: declarations)
+            }.filter { !$0.declarations.isEmpty }
+    }
+    
     public struct Options: OptionsType {
         let files: [String]
         let output: String
@@ -59,14 +67,16 @@ public struct GenerateMocksCommand: CommandType {
         let noTimestamp: Bool
         let testableFrameworks: [String]
         let filePrefix: String
+        let noClassMocking: Bool
         
-        public init(output: String, testableFrameworks: String, noHeader: Bool, noTimestamp: Bool, filePrefix: String, files: [String]) {
+        public init(output: String, testableFrameworks: String, noHeader: Bool, noTimestamp: Bool, filePrefix: String, noClassMocking: Bool, files: [String]) {
             self.output = output
             self.testableFrameworks = testableFrameworks.componentsSeparatedByString(",").filter { !$0.isEmpty }
             self.noHeader = noHeader
             self.noTimestamp = noTimestamp
             self.filePrefix = filePrefix
             self.files = files
+            self.noClassMocking = noClassMocking
         }
         
         public static func evaluate(m: CommandMode) -> Result<Options, CommandantError<CuckooGeneratorError>> {
@@ -76,6 +86,7 @@ public struct GenerateMocksCommand: CommandType {
                 <*> m <| Option(key: "no-header", defaultValue: false, usage: "Do not generate file headers.")
                 <*> m <| Option(key: "no-timestamp", defaultValue: false, usage: "Do not generate timestamp.")
                 <*> m <| Option(key: "file-prefix", defaultValue: "", usage: "Names of generated files in directory will start with this prefix. Only works when output path is directory.")
+                <*> m <| Option(key: "no-class-mocking", defaultValue: false, usage: "Do not generate mocks for classes.")
                 <*> m <| Argument(usage: "Files to parse and generate mocks for.")
         }
     }